Improved unit tests with dbusmock
The upstream verison of python-dbusmock uses incorrect dbus type
definitions for the bluez service objects, so dbus_fast fails the
signature check when comparing the actual values with the type
definitions from the introspection result.

Temporarily switch to a forked version with the fix.

This allows us to add some more robust test assertions about the state
of bluez.