Message protocol: Unpack single values
When encountering a message with the single object marker,
return the value as-is without wrapping it into a tuple container.

This matches the Pybricks behaviour, and partically reverts changes
made in a38028cf1f4a6aa1a31614f9a0e343c5b4e86b19.

However, the virtual hub implementation still wraps single values
into a single object tuple when returning from observer(). This is
due to the type signature on the method of the BLE Hub API, which
does require a tuple to be returned.

This is likely a bug/inaccuracy of the API types, as the Pybricks
firmware itself does not adhere to the type annotation and returns
either a single value, or a tuple.